+++ ext/form/binder.go
@@ -1,4 +1,4 @@
-package xun
+package form
 
 import (
 	"net/http"
