+++ ext/form/validate.go
@@ -1,4 +1,4 @@
-package xun
+package form
 
 import (
 	"github.com/go-playground/locales/en"
